+++ CHANGELOG.md
@@ -12,6 +12,8 @@ The minor version will be incremented upon a breaking change and the patch versi
 
 ### Fixes
 
+- filter: fix memcmp decode data in the config ([#118](https://github.com/lamports-dev/richat/pull/118))
+
 ### Features
 
 - proto: add version to quic response ([#115](https://github.com/lamports-dev/richat/pull/115))